feat(ddm): add multiple widgets
