+++ static/app/views/ddm/ddm.tsx
@@ -1,7 +1,6 @@
 import styled from '@emotion/styled';
 
 import ButtonBar from 'sentry/components/buttonBar';
-import {CompactSelect} from 'sentry/components/compactSelect';
 import EnvironmentPageFilter from 'sentry/components/environmentPageFilter';
 import FeatureBadge from 'sentry/components/featureBadge';
 import {FeatureFeedback} from 'sentry/components/featureFeedback';
@@ -14,14 +13,11 @@ import {PageHeadingQuestionTooltip} from 'sentry/components/pageHeadingQuestionT
 import SentryDocumentTitle from 'sentry/components/sentryDocumentTitle';
 import {t} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
-import {defaultMetricDisplayType, MetricDisplayType} from 'sentry/utils/metrics';
 import useOrganization from 'sentry/utils/useOrganization';
-import useRouter from 'sentry/utils/useRouter';
-import MetricsExplorer from 'sentry/views/ddm/metricsExplorer';
+import MetricsQueryDashboard from 'sentry/views/ddm/metricsQueryDashboard';
 
 function DDM() {
   const organization = useOrganization();
-  const router = useRouter();
 
   return (
     <SentryDocumentTitle title={t('DDM')} orgSlug={organization.slug}>
@@ -52,36 +48,8 @@ function DDM() {
                   <EnvironmentPageFilter />
                   <DatePageFilter />
                 </PageFilterBar>
-                <CompactSelect
-                  triggerProps={{prefix: t('Display')}}
-                  value={router.location.query.display ?? defaultMetricDisplayType}
-                  options={[
-                    {
-                      value: MetricDisplayType.LINE,
-                      label: t('Line Chart'),
-                    },
-                    {
-                      value: MetricDisplayType.AREA,
-                      label: t('Area Chart'),
-                    },
-                    {
-                      value: MetricDisplayType.BAR,
-                      label: t('Bar Chart'),
-                    },
-                  ]}
-                  onChange={({value}) => {
-                    router.push({
-                      ...router.location,
-                      query: {
-                        ...router.location.query,
-                        cursor: undefined,
-                        display: value,
-                      },
-                    });
-                  }}
-                />
               </PaddedContainer>
-              <MetricsExplorer />
+              <MetricsQueryDashboard />
             </Layout.Main>
           </Layout.Body>
         </Layout.Page>
